Created new Robot.java and Subject.java Previous files are termed with the word Old eg: RobotOld.java The old classes are extensions of the new ones
Several files were modified to allow backward compatiblity.
Most loggers now work exclusively with SubjectOld.java and not with Subject.java

ModelAction was renamed to RobotAction and moved to platform along with its subclasses.
The ModelAction port was moved to models.

The new Robot.java is almost empty.
Several Interfaces were created to be used with the Robot.java

A robot Puck.java was created